web: Improved next page check
Greatly optimized the check used to see if a next page is available. It used to check each character for at least two pages which was a big waste of api calls.

Also changed the name of the website exported object to just web.

Made the logger file load the channel in on startup instead of having to load it each and every time it is called.